+++ Sources/code/platform/Models/Order/ShipmentsRequestSchemaOrderPlatformModel.swift
@@ -20,6 +20,8 @@ public extension PlatformClient.Order {
         
         public var dataUpdates: DataUpdates?
         
+        public var transitionComments: [TransitionComments]?
+        
 
         public enum CodingKeys: String, CodingKey {
             
@@ -31,9 +33,11 @@ public extension PlatformClient.Order {
             
             case dataUpdates = "data_updates"
             
+            case transitionComments = "transition_comments"
+            
         }
 
-        public init(dataUpdates: DataUpdates? = nil, identifier: String, products: [Products]? = nil, reasons: ReasonsData? = nil) {
+        public init(dataUpdates: DataUpdates? = nil, identifier: String, products: [Products]? = nil, reasons: ReasonsData? = nil, transitionComments: [TransitionComments]? = nil) {
             
             self.identifier = identifier
             
@@ -42,6 +46,8 @@ public extension PlatformClient.Order {
             self.products = products
             
             self.dataUpdates = dataUpdates
+            
+            self.transitionComments = transitionComments
             
         }
 
@@ -89,6 +95,18 @@ public extension PlatformClient.Order {
                 }
                 
             
+            
+                do {
+                    transitionComments = try container.decode([TransitionComments].self, forKey: .transitionComments)
+                
+                } catch DecodingError.typeMismatch(let type, let context) {
+                    print("Type '\(type)' mismatch:", context.debugDescription)
+                    print("codingPath:", context.codingPath)
+                } catch {
+                    
+                }
+                
+            
         }
         
         public func encode(to encoder: Encoder) throws {
@@ -112,6 +130,11 @@ public extension PlatformClient.Order {
             
             
             try? container.encodeIfPresent(dataUpdates, forKey: .dataUpdates)
+            
+            
+            
+            
+            try? container.encodeIfPresent(transitionComments, forKey: .transitionComments)
             
             
         }
@@ -138,6 +161,8 @@ public extension PlatformClient.ApplicationClient.Order {
         
         public var dataUpdates: DataUpdates?
         
+        public var transitionComments: [TransitionComments]?
+        
 
         public enum CodingKeys: String, CodingKey {
             
@@ -149,9 +174,11 @@ public extension PlatformClient.ApplicationClient.Order {
             
             case dataUpdates = "data_updates"
             
+            case transitionComments = "transition_comments"
+            
         }
 
-        public init(dataUpdates: DataUpdates? = nil, identifier: String, products: [Products]? = nil, reasons: ReasonsData? = nil) {
+        public init(dataUpdates: DataUpdates? = nil, identifier: String, products: [Products]? = nil, reasons: ReasonsData? = nil, transitionComments: [TransitionComments]? = nil) {
             
             self.identifier = identifier
             
@@ -160,6 +187,8 @@ public extension PlatformClient.ApplicationClient.Order {
             self.products = products
             
             self.dataUpdates = dataUpdates
+            
+            self.transitionComments = transitionComments
             
         }
 
@@ -207,6 +236,18 @@ public extension PlatformClient.ApplicationClient.Order {
                 }
                 
             
+            
+                do {
+                    transitionComments = try container.decode([TransitionComments].self, forKey: .transitionComments)
+                
+                } catch DecodingError.typeMismatch(let type, let context) {
+                    print("Type '\(type)' mismatch:", context.debugDescription)
+                    print("codingPath:", context.codingPath)
+                } catch {
+                    
+                }
+                
+            
         }
         
         public func encode(to encoder: Encoder) throws {
@@ -230,6 +271,11 @@ public extension PlatformClient.ApplicationClient.Order {
             
             
             try? container.encodeIfPresent(dataUpdates, forKey: .dataUpdates)
+            
+            
+            
+            
+            try? container.encodeIfPresent(transitionComments, forKey: .transitionComments)
             
             
         }